barriers: share the task bar receiver between x11 and apple
Both of these are mostly noops anyway, apparently only Windows needs
this. So let's share the code.
