+++ src/cmd/barriers/CMakeLists.txt
@@ -25,12 +25,12 @@ if (BUILD_MSWINDOWS)
     set_property(SOURCE barriers.rc PROPERTY COMPILE_FLAGS /nologo)
 endif()
 if (BUILD_CARBON)
-    file(GLOB carbon_headers "OSX*.h")
-    file(GLOB carbon_sources "OSX*.cpp")
+    file(GLOB carbon_headers "OSX*.h" "Fallback*.h")
+    file(GLOB carbon_sources "OSX*.cpp" "Fallback*.cpp")
 endif()
 if (BUILD_XWINDOWS)
-    file(GLOB xwin_headers "XWindows*.h")
-    file(GLOB xwin_sources "XWindows*.cpp")
+    file(GLOB xwin_headers "XWindows*.h" "Fallback*.h")
+    file(GLOB xwin_sources "XWindows*.cpp" "Fallback*.cpp")
 endif()
 
 list(APPEND sources ${mswin_sources} ${carbon_sources} ${xwin_sources})
