+++ src/cmd/barriers/FallbackServerTaskBarReceiver.cpp
@@ -1,7 +1,7 @@
 /*
  * InputLeap -- mouse and keyboard sharing utility
  * Copyright (C) 2012-2016 Symless Ltd.
- * Copyright (C) 2004 Chris Schoeneman
+ * Copyright (C) 2003 Chris Schoeneman
  *
  * This package is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -16,14 +16,14 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "OSXServerTaskBarReceiver.h"
+#include "FallbackServerTaskBarReceiver.h"
 #include "arch/Arch.h"
 
 //
-// OSXServerTaskBarReceiver
+// CFallbackServerTaskBarReceiver
 //
 
-OSXServerTaskBarReceiver::OSXServerTaskBarReceiver(
+CFallbackServerTaskBarReceiver::CFallbackServerTaskBarReceiver(
         const BufferedLogOutputter*, IEventQueue* events) :
     ServerTaskBarReceiver(events)
 {
@@ -31,30 +31,30 @@ OSXServerTaskBarReceiver::OSXServerTaskBarReceiver(
     ARCH->addReceiver(this);
 }
 
-OSXServerTaskBarReceiver::~OSXServerTaskBarReceiver()
+CFallbackServerTaskBarReceiver::~CFallbackServerTaskBarReceiver()
 {
     ARCH->removeReceiver(this);
 }
 
 void
-OSXServerTaskBarReceiver::showStatus()
+CFallbackServerTaskBarReceiver::showStatus()
 {
     // do nothing
 }
 
 void
-OSXServerTaskBarReceiver::runMenu(int, int)
+CFallbackServerTaskBarReceiver::runMenu(int, int)
 {
     // do nothing
 }
 
 void
-OSXServerTaskBarReceiver::primaryAction()
+CFallbackServerTaskBarReceiver::primaryAction()
 {
     // do nothing
 }
 
-IArchTaskBarReceiver::Icon OSXServerTaskBarReceiver::getIcon() const
+IArchTaskBarReceiver::Icon CFallbackServerTaskBarReceiver::getIcon() const
 {
     return nullptr;
 }
@@ -62,5 +62,5 @@ IArchTaskBarReceiver::Icon OSXServerTaskBarReceiver::getIcon() const
 IArchTaskBarReceiver*
 createTaskBarReceiver(const BufferedLogOutputter* logBuffer, IEventQueue* events)
 {
-    return new OSXServerTaskBarReceiver(logBuffer, events);
+    return new CFallbackServerTaskBarReceiver(logBuffer, events);
 }
