--- src/cmd/barriers/XWindowsServerTaskBarReceiver.h
@@ -1,38 +1,0 @@
-/*
- * InputLeap -- mouse and keyboard sharing utility
- * Copyright (C) 2012-2016 Symless Ltd.
- * Copyright (C) 2003 Chris Schoeneman
- *
- * This package is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * found in the file LICENSE that should have accompanied this file.
- *
- * This package is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#pragma once
-
-#include "inputleap/ServerTaskBarReceiver.h"
-
-class BufferedLogOutputter;
-class IEventQueue;
-
-//! Implementation of ServerTaskBarReceiver for X Windows
-class CXWindowsServerTaskBarReceiver : public ServerTaskBarReceiver {
-public:
-    CXWindowsServerTaskBarReceiver(
-        const BufferedLogOutputter*, IEventQueue* events);
-    ~CXWindowsServerTaskBarReceiver() override;
-
-    // IArchTaskBarReceiver overrides
-    void showStatus() override;
-    void runMenu(int x, int y) override;
-    void primaryAction() override;
-    Icon getIcon() const override;
-};
